+++ src/solvers/conjugate_residual.jl
@@ -219,17 +219,17 @@ is initialized with ``X^{(0)}`` as the zero vector and
 
 * the initial residual ``r^{(0)} = -b(p) - $(_tex(:Cal, "A"))(p)[X^{(0)}]``
 * the initial conjugate direction ``d^{(0)} = r^{(0)}``
-* initialize ``Y^{(0)} = $(_tex(:Cal, "A"))(p)[X^{(0)}]``
+* initialize ``Y^{(0)} = $(_tex(:Cal, "A"))(p)[r^{(0)}]``
 
 performed the following steps at iteration ``k=0,…`` until the `stopping_criterion` is fulfilled.
 
 1. compute a step size ``α_k = $(_tex(:displaystyle))$(_tex(:frac, "⟨ r^{(k)}, $(_tex(:Cal, "A"))(p)[r^{(k)}] ⟩_p", "⟨ $(_tex(:Cal, "A"))(p)[d^{(k)}], $(_tex(:Cal, "A"))(p)[d^{(k)}] ⟩_p"))``
 2. do a step ``X^{(k+1)} = X^{(k)} + α_kd^{(k)}``
-2. update the residual ``r^{(k+1)} = r^{(k)} + α_k Y^{(k)}``
+3. update the residual ``r^{(k+1)} = r^{(k)} - α_k Y^{(k)}``
 4. compute ``Z = $(_tex(:Cal, "A"))(p)[r^{(k+1)}]``
 5. Update the conjugate coefficient ``β_k = $(_tex(:displaystyle))$(_tex(:frac, "⟨ r^{(k+1)}, $(_tex(:Cal, "A"))(p)[r^{(k+1)}] ⟩_p", "⟨ r^{(k)}, $(_tex(:Cal, "A"))(p)[r^{(k)}] ⟩_p"))``
 6. Update the conjugate direction ``d^{(k+1)} = r^{(k+1)} + β_kd^{(k)}``
-7. Update  ``Y^{(k+1)} = -Z + β_k Y^{(k)}``
+7. Update  ``Y^{(k+1)} = Z + β_k Y^{(k)}``
 
 Note that the right hand side of Step 7 is the same as evaluating ``$(_tex(:Cal, "A"))[d^{(k+1)}]``, but avoids the actual evaluation